refactor: update getCurrentTime to return Int and adjust related code in AudioModule

+++ packages/audio/ios/Extension/SongModelExtension.swift
@@ -8,8 +8,8 @@ extension SongModelContract {
           "artistId": String(self.artistId.raw),
           "artistName": self.artistName,
           "imageId": self.imageId,
-          "duration": self.duration,
-          "trackNumber": Double(self.trackNumber)
+          "duration": Int(self.duration),
+          "trackNumber": self.trackNumber
         ]
     }
 }
